(#72): 모든 경로에 따른 크기 조절 기능 추가
